+++ mirrord/operator/src/client/database_branches.rs
@@ -337,6 +337,7 @@ impl DatabaseBranchParams {
                     let params = PgBranchParams::new(id.as_ref(), pg_config, target);
                     pg.insert(id, params);
                 }
+                DatabaseBranchConfig::Redis(_) => {}
             };
         }
         Self { mysql, pg }
@@ -408,6 +409,13 @@ impl MysqlBranchParams {
                     container: container.clone(),
                     variable: variable.clone(),
                 }),
+                ConnectionSourceKind::EnvFrom {
+                    container,
+                    variable,
+                } => CrdConnectionSourceMysql::Url(CrdConnectionSourceKindMysql::EnvFrom {
+                    container: container.clone(),
+                    variable: variable.clone(),
+                }),
             },
         };
         let spec = MysqlBranchDatabaseSpec {
@@ -447,6 +455,13 @@ impl PgBranchParams {
                     container,
                     variable,
                 } => CrdConnectionSourcePg::Url(CrdConnectionSourceKindPg::Env {
+                    container: container.clone(),
+                    variable: variable.clone(),
+                }),
+                ConnectionSourceKind::EnvFrom {
+                    container,
+                    variable,
+                } => CrdConnectionSourcePg::Url(CrdConnectionSourceKindPg::EnvFrom {
                     container: container.clone(),
                     variable: variable.clone(),
                 }),